+++ src/utils.rs
@@ -4,7 +4,7 @@ use reqwest;
 use reqwest::blocking::Client;
 use reqwest::header::USER_AGENT;
 use serde_json;
-use std::process;
+use std::{env, fs, process};
 
 const VERSION: &str = env!("CARGO_PKG_VERSION");
 
@@ -18,6 +18,33 @@ macro_rules! debug_log {
     };
 }
 
+fn is_systemd_present() {
+    match process::Command::new("ps")
+        .arg("-p")
+        .arg("1")
+        .arg("-o")
+        .arg("comm=")
+        .output()
+    {
+        Ok(output) => {
+            let init_process = String::from_utf8_lossy(&output.stdout);
+            if init_process.trim() == "systemd" {
+                return;
+            }
+            println!("\x1b[33;1mINFO: Most likely systemd is not available on your system.\x1b[0m");
+        }
+        Err(_) => {
+            println!(
+                "\x1b[33;1mINFO: Could not detect if systemd is available on your system.\x1b[0m"
+            )
+        }
+    }
+    println!(
+        "You can try using XDG Autostart instead to add the application to autostart without systemd."
+    );
+    println!("Use the \x1b[32;1m--xdg\x1b[0m flag with the subcommands like this: \x1b[32;1mmpris-discord-rpc enable --xdg\x1b[0m.");
+}
+
 pub fn enable_service() {
     match process::Command::new("systemctl")
         .arg("--user")
@@ -27,6 +54,7 @@ pub fn enable_service() {
         Ok(_) => println!("Reloaded user systemd services."),
         Err(_) => {
             println!("Failed to reload user systemd services.");
+            is_systemd_present();
             process::exit(1);
         }
     }
@@ -34,25 +62,14 @@ pub fn enable_service() {
     match process::Command::new("systemctl")
         .arg("--user")
         .arg("enable")
+        .arg("--now")
         .arg("mpris-discord-rpc.service")
         .status()
     {
-        Ok(_) => println!("Enabled user systemd service."),
+        Ok(_) => println!("Enabled and started user systemd service."),
         Err(_) => {
-            println!("Failed to enable user systemd service.");
-            process::exit(1);
-        }
-    }
-
-    match process::Command::new("systemctl")
-        .arg("--user")
-        .arg("start")
-        .arg("mpris-discord-rpc.service")
-        .status()
-    {
-        Ok(_) => println!("Started user systemd service."),
-        Err(_) => {
-            println!("Failed to start user systemd service.");
+            println!("Failed to enable and start user systemd service.");
+            is_systemd_present();
             process::exit(1);
         }
     }
@@ -63,26 +80,15 @@ pub fn enable_service() {
 pub fn disable_service() {
     match process::Command::new("systemctl")
         .arg("--user")
-        .arg("stop")
-        .arg("mpris-discord-rpc.service")
-        .status()
-    {
-        Ok(_) => println!("Stopped user systemd service."),
-        Err(_) => {
-            println!("Failed to stop user systemd service.");
-            process::exit(1);
-        }
-    }
-
-    match process::Command::new("systemctl")
-        .arg("--user")
         .arg("disable")
+        .arg("--now")
         .arg("mpris-discord-rpc.service")
         .status()
     {
-        Ok(_) => println!("Disabled user systemd service."),
+        Ok(_) => println!("Stopped and disabled user systemd service."),
         Err(_) => {
-            println!("Failed to disable user systemd service.");
+            println!("Failed to stop and disable user systemd service.");
+            is_systemd_present();
             process::exit(1);
         }
     }
@@ -103,6 +109,81 @@ pub fn restart_service() {
             process::exit(1);
         }
     }
+    process::exit(0);
+}
+
+pub fn get_config_path() -> Option<std::path::PathBuf> {
+    if let Some(config_home) = env::var_os("XDG_CONFIG_HOME") {
+        Some(std::path::PathBuf::from(config_home))
+    } else if let Some(home_dir) = env::var_os("HOME") {
+        let mut path = std::path::PathBuf::from(home_dir);
+        path.push(".config");
+        Some(path)
+    } else {
+        None
+    }
+}
+
+pub fn add_xdg_autostart() {
+    let mut desktopt_file_path = match get_config_path() {
+        Some(path) => path,
+        None => {
+            println!("\x1b[31mWARNING: Failed to determine user config directory.\x1b[0m");
+            process::exit(1);
+        }
+    };
+    desktopt_file_path.push("autostart");
+    desktopt_file_path.push("mpris-discord-rpc.desktop");
+
+    let desktop_file_content = r#"[Desktop Entry]
+Name=mpris-discord-rpc
+Type=Application
+Exec=mpris-discord-rpc
+X-GNOME-Autostart-enabled=true
+Hidden=false
+StartupNotify=false
+Terminal=false
+"#;
+
+    match fs::write(&desktopt_file_path, desktop_file_content) {
+        Ok(_) => {
+            println!(
+                "Created file: \x1b[32;1m{}\x1b[0m ",
+                desktopt_file_path.display()
+            );
+            println!("This RPC should now start automatically with your system if your DE/WM supports XDG Autostart.");
+        }
+        Err(_) => {
+            println!("\x1b[31mERROR: Failed to create autostart .desktop file.\x1b[0m");
+            process::exit(1);
+        }
+    }
+
+    process::exit(0);
+}
+
+pub fn remove_xdg_autostart() {
+    let mut desktopt_file_path = match get_config_path() {
+        Some(path) => path,
+        None => {
+            println!("\x1b[31mWARNING: Failed to determine user config directory.\x1b[0m");
+            process::exit(1);
+        }
+    };
+    desktopt_file_path.push("autostart");
+    desktopt_file_path.push("mpris-discord-rpc.desktop");
+
+    match fs::remove_file(&desktopt_file_path) {
+        Ok(_) => println!(
+            "Removed file: \x1b[32;1m{}\x1b[0m ",
+            desktopt_file_path.display()
+        ),
+        Err(_) => {
+            println!("\x1b[31mERROR: Failed to remove autostart .desktop file.\x1b[0m");
+            process::exit(1);
+        }
+    }
+
     process::exit(0);
 }
 
@@ -326,9 +407,3 @@ pub fn sanitize_name(input: &str) -> String {
         .filter(|c| c.is_alphanumeric() || *c == '_')
         .collect()
 }
-
-// pub fn format_time(seconds: u64) -> String {
-//     let minutes = seconds / 60;
-//     let remaining_seconds = seconds % 60;
-//     format!("{:02}:{:02}", minutes, remaining_seconds)
-// }